requirements.txt
groq
PyMuPDF
python-dotenv
streamlit
python-docx
